+++ src/collectors/mod.rs
@@ -2,5 +2,6 @@ pub mod diskstats;
 #[cfg(test)]
 pub mod iostat;
 pub mod md_sysfs;
+#[cfg(test)]
 pub mod raid;
 pub mod smart;